+++ pkg/cmd/get/clusterclaim/exec.go
@@ -2,12 +2,15 @@
 package clusterclaim
 
 import (
+	"context"
 	"fmt"
 
 	printclusterpoolv1alpha1 "github.com/stolostron/cm-cli/api/cm-cli/v1alpha1"
 	"github.com/stolostron/cm-cli/pkg/clusterpoolhost"
 	"github.com/stolostron/cm-cli/pkg/helpers"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/client-go/kubernetes"
 
 	"github.com/spf13/cobra"
 )
@@ -37,10 +40,44 @@ func (o *Options) run() (err error) {
 	if len(o.ClusterClaim) == 0 {
 		err = o.getCCS(cphs)
 	} else {
-		err = o.getCC(cph)
+		switch o.KubeConfig {
+		case true:
+			err = o.getKubeConfig(cph)
+		case false:
+			err = o.getCC(cph)
+		}
 	}
 	return err
 
+}
+
+func (o *Options) getKubeConfig(cph *clusterpoolhost.ClusterPoolHost) (err error) {
+	cc, err := cph.GetClusterClaim(o.ClusterClaim, o.WithCredentials, o.Timeout, o.CMFlags.DryRun, o.GetOptions.PrintFlags)
+	if err != nil {
+		return err
+	}
+	cd, err := cph.GetClusterDeployment(cc)
+	if err != nil {
+		return err
+	}
+	kubeConfigSecretName := cd.Spec.ClusterMetadata.AdminKubeconfigSecretRef.Name
+	clusterPoolRestConfig, err := cph.GetGlobalRestConfig()
+	if err != nil {
+		return err
+	}
+	kubeClient, err := kubernetes.NewForConfig(clusterPoolRestConfig)
+	if err != nil {
+		return err
+	}
+	kubeConfigSecret, err := kubeClient.
+		CoreV1().
+		Secrets(cd.Namespace).
+		Get(context.TODO(), kubeConfigSecretName, metav1.GetOptions{})
+	if err != nil {
+		return err
+	}
+	fmt.Println(string(kubeConfigSecret.Data["kubeconfig"]))
+	return nil
 }
 
 func (o *Options) getCC(cph *clusterpoolhost.ClusterPoolHost) (err error) {