+++ pkg/helpers/client.go
@@ -25,4 +25,5 @@ var (
 	GvrIDPConfig                schema.GroupVersionResource = schema.GroupVersionResource{Group: "identityconfig.identitatem.io", Version: "v1alpha1", Resource: "idpconfigs"}
 	GvrOpenshiftClusterVersions schema.GroupVersionResource = schema.GroupVersionResource{Group: "config.openshift.io", Version: "v1", Resource: "clusterversions"}
 	GvrHC                       schema.GroupVersionResource = schema.GroupVersionResource{Group: "hypershift.openshift.io", Version: "v1alpha1", Resource: "hostedclusters"}
+	GvrHD                       schema.GroupVersionResource = schema.GroupVersionResource{Group: "cluster.open-cluster-management.io", Version: "v1alpha1", Resource: "hypershiftdeployments"}
 )